--- src/pages/VersionPage/CreateVersionPage/CreateVersionPage.scss
@@ -1,26 +1,0 @@
-.form-create{
-    background: white;
-    margin: 0px auto;
-    border-radius: 10px;
-    margin-top: 180px;
-    max-width:70%;
-    .form-header{
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        border-bottom: 1px solid rgba(0, 0, 0, .05);
-        padding:30px 20px ;
-        &__control{
-            min-width: 20%;
-            display: flex;
-            justify-content: space-around;
-        }
-    }
-    .form-body{
-        min-width: 100%;
-        display: flex;
-        justify-content:space-around;
-        align-items: center;
-        padding:30px; 
-    }
-}
