+++ src/pages/VersionPage/VersionPage.scss
@@ -2,7 +2,7 @@
 .version-page {
     margin: 30px;
     margin-top: 200px;
-  .form-version {
+  .filter-version {
     border-radius: 10px;
     background-color:white;
     margin: 30px auto;
