+++ src/pages/VersionPage/VersionPage.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
+import { Link } from "react-router-dom";
 import { Button } from "../../components";
 import { apiVersion } from "../../services/api/apiVersion";
 import { FilterVersions, TableVersions } from "./components";
 import "./VersionPage.scss";
-
 
 export const VersionPage = () => {
   const [dataVersions, setDataVersions] = useState<Array<IResVersion>>([]);
@@ -13,18 +13,19 @@ export const VersionPage = () => {
         .getListVersions({})
         .then((res) => res.data)
         .then((data: IResGetListVersions) => {
-          setDataVersions(data.Content.Versions);
+          const listVersions = data.Content.Versions;
+          setDataVersions(listVersions);
         });
     } catch (error) {
       console.log(error);
     }
   }, []);
-  const handleSubmit = (searchVersion: string, searchStatus: number) => {
+  const handleSearch = (Title: string, Status: number) => {
     try {
       apiVersion
         .getListVersions({
-          Title: searchVersion,
-          Status: searchStatus,
+          Title,
+          Status,
         })
         .then((res) => res.data)
         .then((data: IResGetListVersions) => {
@@ -33,20 +34,20 @@ export const VersionPage = () => {
     } catch (error) {
       console.log(error);
     }
-  }
-  const handleCreate = () => {
-    
-  }
+  };
+  const handleCreate = () => {};
 
   return (
     <div className="version-page">
-      <div className="form-version">
-        <FilterVersions handleSubmit={handleSubmit} />
+      <div className="filter-version">
+        <FilterVersions handleSubmit={handleSearch} />
       </div>
       <div className="list-version">
         <div className="list-version__header">
           <h3>Danh Sách Version</h3>
-          <Button isCreate type="submit"  handleOnClick = {handleCreate}  />
+          <Link to={"/version/create"}>
+            <Button isCreate type="submit" handleOnClick={handleCreate} />
+          </Link>
         </div>
         <TableVersions dataVersions={dataVersions} />
       </div>
