--- src/pages/VersionPage/components/FilterVersion/FilterVersion.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import "./FilterVersion.scss";
-
-export const FilterVersion = () => {
-  return <div>FilterTasks</div>;
-};
