+++ src/pages/VersionPage/components/FilterVersions/FilterVersions.scss
@@ -1,4 +1,4 @@
-.form-filter{
+.form-version{
     .form-header{
         display: flex;
         justify-content: space-between;
