--- src/pages/VersionPage/components/TableVersion/TableVersion.tsx
@@ -1,37 +1,0 @@
-import React, { FC, Fragment } from "react";
-import { Actions, Table } from "../../../../components";
-
-export const TableVersion: FC<ITableVersion> = ({ dataVersions }) => {
-  return (
-    <Table
-      thead={["Tiêu đề", "Trạng thái", "Thông tin tạo", "Thông tin sửa"]}
-      tbody={
-        <Fragment>
-          {dataVersions.map((info) => (
-            <tr>
-              <td>
-                <b>{info.Title}</b>
-              </td>
-              <td>
-                <b>{info.StatusText}</b>
-              </td>
-              <td>
-                {info.CreateUserName}
-                <br />
-                {info.CreateDate}
-              </td>
-              <td>
-                {info.UpdateUserName}
-                <br />
-                {info.UpdateDate}
-              </td>
-              <td>
-                <Actions />
-              </td>
-            </tr>
-          ))}
-        </Fragment>
-      }
-    ></Table>
-  );
-};
